Fixed a unsigned int bug. Still have compile warnings (but it compiles)

+++ aimbots-src/src/subsystems/hud_display/hud_constants.hpp
@@ -5,8 +5,8 @@
 
 #include "modm/processing/resumable.hpp"
 
-static constexpr uint8_t SCREEN_WIDTH = 1920;
-static constexpr uint8_t SCREEN_HEIGHT = 1080;
+static constexpr uint16_t SCREEN_WIDTH = 1920;
+static constexpr uint16_t SCREEN_HEIGHT = 1080;
 
 static constexpr uint8_t DEFAULT_GRAPHIC_LAYER = 0;
